fix(tests): EIP-7702: Move invalid tx tests to its own file
